requirements.txt
langchain==0.1.0
langchain-community==0.0.10
langchain-huggingface==0.0.1
langchain-google-genai==1.0.3  # For Gemini
google-generativeai==0.4.0  # Gemini SDK
faiss-cpu==1.9.0.post1
sentence-transformers==2.2.2
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pandas==2.1.3
numpy==1.24.3
xarray==2023.10.0
fuzzywuzzy==0.18.0
python-dotenv==1.0.0
loguru==0.7.2
slowapi==0.1.9 # For rate limiting (was missing in prev)